% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{anova}
\alias{anova}
\title{levene's test en anova}
\usage{
anova(x, y, data = NULL)
}
\arguments{
\item{x}{- onafhankelijke variabele}

\item{y}{- afhankeleijke variabele}

\item{data}{- database}
}
\value{
levene's test en anova
}
\description{
levene's test en anova
}
