% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{syntax.plot}
\alias{syntax.plot}
\title{syntax voor het plotten van verschillende figuren}
\usage{
syntax.plot()
}
\value{
a text
}
\description{
syntax voor het plotten van verschillende figuren
}
\examples{
syntax.plot()
}
