% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntax.,regressie.R, R/syntax.t.toets.R
\name{syntax.t.toets}
\alias{syntax.t.toets}
\title{syntax voor t-toets}
\usage{
syntax.t.toets()

syntax.t.toets()
}
\value{
a text

a text
}
\description{
syntax voor t-toets

syntax voor t-toets
}
\examples{
syntax.t.toets()
syntax.t.toets()
}
